+++ fsutil/fs_test.go
@@ -321,6 +321,7 @@ func (s *FSSuite) TestGetTime(c *check.C) {
 	ats, mts, cts, err = GetTimestamps("/not_exist")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.ErrorMatches, `Can't get file info for "/not_exist": no such file or directory`)
 	c.Assert(ats, check.Equals, int64(-1))
 	c.Assert(mts, check.Equals, int64(-1))
 	c.Assert(cts, check.Equals, int64(-1))
@@ -343,16 +344,19 @@ func (s *FSSuite) TestGetTime(c *check.C) {
 	at, err = GetATime("")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.Equals, ErrEmptyPath)
 	c.Assert(at.IsZero(), check.Equals, true)
 
 	mt, err = GetMTime("")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.Equals, ErrEmptyPath)
 	c.Assert(mt.IsZero(), check.Equals, true)
 
 	ct, err = GetCTime("")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.Equals, ErrEmptyPath)
 	c.Assert(ct.IsZero(), check.Equals, true)
 }
 
@@ -378,6 +382,7 @@ func (s *FSSuite) TestGetOwner(c *check.C) {
 	uid, gid, err = GetOwner("/not_exist")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.ErrorMatches, `Can't get owner info for "/not_exist": no such file or directory`)
 	c.Assert(uid, check.Equals, -1)
 	c.Assert(gid, check.Equals, -1)
 }
@@ -633,11 +638,13 @@ func (s *FSSuite) TestCountLines(c *check.C) {
 	n, err := CountLines("")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.Equals, ErrEmptyPath)
 	c.Assert(n, check.Equals, 0)
 
 	n, err = CountLines("/not_exist")
 
 	c.Assert(err, check.NotNil)
+	c.Assert(err, check.ErrorMatches, `open /not_exist: no such file or directory`)
 	c.Assert(n, check.Equals, 0)
 
 	n, err = CountLines(tmpFile)
@@ -661,14 +668,14 @@ func (s *FSSuite) TestCopyFile(c *check.C) {
 	os.Chmod(tmpFile3, 0111)
 	os.Chmod(tmpDir3, 0500)
 
-	c.Assert(CopyFile("", tmpFile2), check.NotNil)
-	c.Assert(CopyFile(tmpFile1, ""), check.NotNil)
-	c.Assert(CopyFile("/not_exist", tmpFile2), check.NotNil)
-	c.Assert(CopyFile(tmpDir1, tmpFile2), check.NotNil)
-	c.Assert(CopyFile(tmpFile3, tmpFile2), check.NotNil)
-	c.Assert(CopyFile(tmpFile1, "/not_exist/test.file"), check.NotNil)
-	c.Assert(CopyFile(tmpFile1, tmpDir3+"/test.file"), check.NotNil)
-	c.Assert(CopyFile(tmpFile1, tmpFile3), check.NotNil)
+	c.Assert(CopyFile("", tmpFile2), check.ErrorMatches, `Can't copy file: Source file can't be blank`)
+	c.Assert(CopyFile(tmpFile1, ""), check.ErrorMatches, `Can't copy file: Target file can't be blank`)
+	c.Assert(CopyFile("/not_exist", tmpFile2), check.ErrorMatches, `Can't copy file: File /not_exist does not exists`)
+	c.Assert(CopyFile(tmpDir1, tmpFile2), check.ErrorMatches, `Can't copy file: File /tmp/check-[0-9]+/[0-9]+ is not a regular file`)
+	c.Assert(CopyFile(tmpFile3, tmpFile2), check.ErrorMatches, `Can't copy file: File /tmp/check-[0-9]+/[0-9]+/test3.file is not readable`)
+	c.Assert(CopyFile(tmpFile1, "/not_exist/test.file"), check.ErrorMatches, `Can't copy file: Directory /not_exist does not exists`)
+	c.Assert(CopyFile(tmpFile1, tmpDir3+"/test.file"), check.ErrorMatches, `Can't copy file: Directory /tmp/check-[0-9]+/[0-9]+ is not writable`)
+	c.Assert(CopyFile(tmpFile1, tmpFile3), check.ErrorMatches, `Can't copy file: Target file /tmp/check-[0-9]+/[0-9]+/test3.file is not writable`)
 
 	c.Assert(CopyFile(tmpFile1, tmpFile2, 0600), check.IsNil)
 	c.Assert(GetSize(tmpFile2), check.Equals, int64(5))
@@ -684,8 +691,8 @@ func (s *FSSuite) TestCopyFile(c *check.C) {
 
 	_disableCopyFileChecks = true
 
-	c.Assert(CopyFile("", tmpFile2), check.NotNil)
-	c.Assert(CopyFile(tmpFile1, ""), check.NotNil)
+	c.Assert(CopyFile("", tmpFile2), check.ErrorMatches, `read .: is a directory`)
+	c.Assert(CopyFile(tmpFile1, ""), check.ErrorMatches, `open .: is a directory`)
 }
 
 func (s *FSSuite) TestMoveFile(c *check.C) {
@@ -704,13 +711,13 @@ func (s *FSSuite) TestMoveFile(c *check.C) {
 	os.Chmod(tmpFile3, 0111)
 	os.Chmod(tmpDir2, 0500)
 
-	c.Assert(MoveFile("", tmpFile2), check.NotNil)
-	c.Assert(MoveFile(tmpFile1, ""), check.NotNil)
-	c.Assert(MoveFile("/not_exist", tmpFile2), check.NotNil)
-	c.Assert(MoveFile(tmpDir, tmpFile2), check.NotNil)
-	c.Assert(MoveFile(tmpFile3, tmpFile2), check.NotNil)
-	c.Assert(MoveFile(tmpFile1, "/not_exist/file.test"), check.NotNil)
-	c.Assert(MoveFile(tmpFile1, tmpDir2+"/file.test"), check.NotNil)
+	c.Assert(MoveFile("", tmpFile2), check.ErrorMatches, `Can't move file: Source file can't be blank`)
+	c.Assert(MoveFile(tmpFile1, ""), check.ErrorMatches, `Can't move file: Target file can't be blank`)
+	c.Assert(MoveFile("/not_exist", tmpFile2), check.ErrorMatches, `Can't move file: File /not_exist does not exists`)
+	c.Assert(MoveFile(tmpDir, tmpFile2), check.ErrorMatches, `Can't move file: File /tmp/check-[0-9]+/[0-9]+ is not a regular file`)
+	c.Assert(MoveFile(tmpFile3, tmpFile2), check.ErrorMatches, `Can't move file: File /tmp/check-[0-9]+/[0-9]+/test3.file is not readable`)
+	c.Assert(MoveFile(tmpFile1, "/not_exist/file.test"), check.ErrorMatches, `Can't move file: Directory /not_exist does not exists`)
+	c.Assert(MoveFile(tmpFile1, tmpDir2+"/file.test"), check.ErrorMatches, `Can't move file: Directory /tmp/check-[0-9]+/[0-9]+ is not writable`)
 
 	c.Assert(MoveFile(tmpFile1, tmpFile2), check.IsNil)
 	c.Assert(MoveFile(tmpFile2, tmpFile1, 0600), check.IsNil)
@@ -757,13 +764,13 @@ func (s *FSSuite) TestCopyDir(c *check.C) {
 
 	c.Assert(list1, check.DeepEquals, list2)
 
-	c.Assert(CopyDir("", targetDir), check.NotNil)
-	c.Assert(CopyDir(sourceDir, ""), check.NotNil)
-	c.Assert(CopyDir(sourceDir+"1", targetDir), check.NotNil)
-	c.Assert(CopyDir(tmpFile1, targetDir), check.NotNil)
-	c.Assert(CopyDir(tmpDir5, targetDir), check.NotNil)
-	c.Assert(CopyDir(tmpDir1, tmpFile2), check.NotNil)
-	c.Assert(CopyDir(tmpDir1, tmpDir6), check.NotNil)
+	c.Assert(CopyDir("", targetDir), check.ErrorMatches, `Can't copy directory: Source directory can't be blank`)
+	c.Assert(CopyDir(sourceDir, ""), check.ErrorMatches, `Can't copy directory: Target directory can't be blank`)
+	c.Assert(CopyDir(sourceDir+"1", targetDir), check.ErrorMatches, `Can't copy directory: Directory /tmp/check-[0-9]+/[0-9]+ does not exists`)
+	c.Assert(CopyDir(tmpFile1, targetDir), check.ErrorMatches, `Can't copy directory: Target /tmp/check-[0-9]+/[0-9]+/test1.file is not a directory`)
+	c.Assert(CopyDir(tmpDir5, targetDir), check.ErrorMatches, `Can't copy directory: Directory /tmp/check-[0-9]+/[0-9]+/test5 is not readable`)
+	c.Assert(CopyDir(tmpDir1, tmpFile2), check.ErrorMatches, `Can't copy directory: Target /tmp/check-[0-9]+/[0-9]+/test2/test2.file is not a directory`)
+	c.Assert(CopyDir(tmpDir1, tmpDir6), check.ErrorMatches, `Can't copy directory: Directory /tmp/check-[0-9]+/[0-9]+/test6 is not writable`)
 
 	_disableCopyDirChecks = true